Former-commit-id: 245597e100b87a4a86e30570665e1e5468291184
Former-commit-id: 75399a023a0d28a4bb4c5e0e302a39712ed6c889
Former-commit-id: 063368343bece83e0196180f362ab661a3e7117d

+++ Frame/SDK/KernelPlugin/AFCClassModule.cpp
@@ -253,14 +253,13 @@ bool AFCClassModule::AddClassInclude(const char* pstrClassFilePath, ARK_SHARE_PT
 
     //////////////////////////////////////////////////////////////////////////
     rapidxml::xml_document<> xDoc;
-    int nDataSize = 0;
+    size_t nDataSize = 0;
 
     std::string strFile = pPluginManager->GetConfigPath() + pstrClassFilePath;
     rapidxml::file<> fdoc(strFile.c_str());
     nDataSize = fdoc.size();
     ARK_SHARE_PTR<char> pData(new char[nDataSize + 1]);
     strncpy(pData.get(), fdoc.data(), nDataSize);
-
 
     pData.get()[nDataSize] = 0;
     xDoc.parse<0>(pData.get());
